Move AppButton component to elements dir
